Rewrite fee calculation code
Remove seekbar because it was not really useful. The fee on the
transaction is now correctly computed. Before it was done after the
profit margins were calculated, which was a mistake. Moved and
refactored the fee percentage variable as a private instance variable.
Not sure why I declared it differently before.